State: Move Jetpack module notices to the respective actions
