+++ client/state/jetpack/modules/actions.js
@@ -2,10 +2,12 @@
  * External dependencies
  */
 import { omit, mapValues } from 'lodash';
+import { translate } from 'i18n-calypso';
 
 /**
  * Internal dependencies
  */
+import { errorNotice, successNotice } from 'calypso/state/notices/actions';
 import {
 	JETPACK_MODULE_ACTIVATE,
 	JETPACK_MODULE_ACTIVATE_FAILURE,
@@ -21,6 +23,34 @@ import {
 import wp from 'calypso/lib/wp';
 
 import 'calypso/state/jetpack/init';
+
+const noticeSettings = {
+	duration: 10000,
+	id: 'site-settings-save',
+};
+const showNotice = ( type, moduleSlug, silent ) => {
+	return ( dispatch ) => {
+		if ( silent ) {
+			return;
+		}
+
+		switch ( type ) {
+			case JETPACK_MODULE_ACTIVATE_SUCCESS:
+			case JETPACK_MODULE_DEACTIVATE_SUCCESS:
+				return dispatch(
+					successNotice( translate( 'Settings saved successfully!' ), noticeSettings )
+				);
+			case JETPACK_MODULE_ACTIVATE_FAILURE:
+			case JETPACK_MODULE_DEACTIVATE_FAILURE:
+				return dispatch(
+					errorNotice(
+						translate( 'There was a problem saving your changes. Please try again.' ),
+						noticeSettings
+					)
+				);
+		}
+	};
+};
 
 export const activateModule = ( siteId, moduleSlug, silent = false ) => {
 	return ( dispatch ) => {
@@ -41,6 +71,7 @@ export const activateModule = ( siteId, moduleSlug, silent = false ) => {
 					moduleSlug,
 					silent,
 				} );
+				dispatch( showNotice( JETPACK_MODULE_ACTIVATE_SUCCESS, moduleSlug, silent ) );
 			} )
 			.catch( ( error ) => {
 				dispatch( {
@@ -50,6 +81,7 @@ export const activateModule = ( siteId, moduleSlug, silent = false ) => {
 					silent,
 					error: error.message,
 				} );
+				dispatch( showNotice( JETPACK_MODULE_ACTIVATE_FAILURE, moduleSlug, silent ) );
 			} );
 	};
 };
@@ -73,6 +105,7 @@ export const deactivateModule = ( siteId, moduleSlug, silent = false ) => {
 					moduleSlug,
 					silent,
 				} );
+				dispatch( showNotice( JETPACK_MODULE_DEACTIVATE_SUCCESS, moduleSlug, silent ) );
 			} )
 			.catch( ( error ) => {
 				dispatch( {
@@ -82,6 +115,7 @@ export const deactivateModule = ( siteId, moduleSlug, silent = false ) => {
 					silent,
 					error: error.message,
 				} );
+				dispatch( showNotice( JETPACK_MODULE_DEACTIVATE_FAILURE, moduleSlug, silent ) );
 			} );
 	};
 };
